backend/requirements.txt
fastapi==0.135.1
openai==2.29.0
pydantic==2.12.5
python-dotenv==1.2.2
SQLAlchemy==2.0.48
starlette==0.52.1
uvicorn
psycopg2-binary